ARM: dts: msm: Add cci gpio resources for vmrm
Due to cci is using pinctrl, but it does not define gpio info
and vmrm is checking the gpios parameter to get the device use
gpio info, so need to add gpio for cci. And due to cci is using
pinctrl, so there is not need to define gpio-req-tbl-num for cci,
so cci just has gpios without gpio-req-tbl-num. This commit add
gpio_for_vmrm_purpose to identify the scene and add gpio resources.

CRs-Fixed: 3685452
Change-Id: I75450c252c62899c6b563745ef3322520595a8eb
Signed-off-by: zhuo <quic_zhuo@quicinc.com>