ARM: dts: msm: Vote MMNOC BW for cam_cc_titan_top_gdsc on Sun
The cam_cc_titan_top_gdsc causes camera to permanently assert qactive
when it's enabled. If MMNOC attempts to collapse when the GDSC is
enabled, then the qactive signal will cause BCM to get stuck in the
collapse sequence. This will eventually lead to rpmh driver timeouts
when requested for MMNOC. Fix this by explicitly voting for MMNOC on
behalf of the GDSC.

Change-Id: I6689318f48fbdb4bd70b192143a11148acdb456e
Signed-off-by: Vivek Aknurwar <quic_viveka@quicinc.com>