ARM: dts: msm: add primary SPMI Arbiter and SPMI debug bus for tuna boards
Add spmi-pmic-arb devices for the primary and secondary SPMI buses
found on tuna.  The primary bus operates at 19.2 MHz and is used
for most of the PMICs. The secondary bus operates at 4.8 MHz
and is used exclusively for charging PMICs. Note that the
secondary bus is not used so it is kept disabled.

Add SPMI debug device and associated child devices for the primary
SPMI interface. This provides consumers with unrestricted access
to the PMIC registers on pre-production devices. This helps
make debugging easier.

Change-Id: I9efadb5082389a519f76f7b5db43f0bde84f2239
Signed-off-by: Kavya Nunna <quic_knunna@quicinc.com>