ARM: dts: msm: shutdown as pinctrl state is missing for UART node in parrot
As part of initial commit '57abd8093413f0969d6c14d38028486e0d303e74
(ARM: dts: msm: Add QUPv3 and GPI DT nodes on Parrot")',
pinctrl state for "shutdown" is missing because of which
we are not configuring the pinctrl state when shutdown
pinctrl state is being  called.

Add "shutdown" pinctrl state in parrot for uart node.

Change-Id: I9d04f6c649652918db0198069c0fd2d19221590b
Signed-off-by: Prasanna S <quic_prass@quicinc.com>