ARM: dts: msm: Vote MMNOC BW for cam_cc_titan_top_gdsc on Tuna
The cam_cc_titan_top_gdsc causes camera to permanently assert qactive
when it's enabled. If MMNOC attempts to collapse when the GDSC is
enabled, then the qactive signal will cause BCM to get stuck in the
collapse sequence. This will eventually lead to rpmh driver timeouts
when requested for MMNOC. Fix this by explicitly voting for MMNOC on
behalf of the GDSC.

Change-Id: I6e75a8da8867a64389fec07ad8ca506b6ac3e860
Signed-off-by: Ajit Pandey <quic_ajipan@quicinc.com>