ARM: dts: msm: Fix trust_ui_vm_mem alignment for Parrot
Parrot need to follow non-relocatable absolute addresses for VM due to
firmware constraints. VM kernel load address is the start address of the
vm_mem_region and build system requires the load address of the kernel
to be 2MB aligned.

This patch adjusts the start address of the trust_ui_vm_mem region for
Parrot to align to 2MB. This also means reducing its size by 1MB.

Change-Id: I1816ec5f6ff18f55ecc4dec39958d442151f8cb0
Signed-off-by: Hrishabh Rajput <quic_hrishabh@quicinc.com>
Signed-off-by: Swetha Chikkaboraiah <quic_schikk@quicinc.com>