ARM: dts: msm: Add GMU CX GenPD instance
Currently, there is a race condition in GenPD framework where
GPU CX GDSC can remain ON if both GMU and KGSL SMMU devices are
suspending in parallel and are voting on the same power domain.
Guidance from genpd team is to use a dedicated power domain for
CX GDSC voting.

Change-Id: Iffeb9a7f24a5e3c31a325e57b021f87f8f94c7fb
Signed-off-by: Kamal Agrawal <quic_kamaagra@quicinc.com>