ARM: dts: msm: add qcom,dsi-select-sec-sync-clocks property
Currently, we are using the qcom,dsi-select-sec-clocks
property to select the clock source for secondary display
when sync mode is enabled.

In the use case, when DPU0 is faulty and DPU1 is working,
the clock source of the secondary display needs to be changed
from _mclk to _clk2.

Add a new property called qcom,dsi-select-sec-sync-clocks to
specify the clock source of the secondary display when running
in sync mode with dual DPU. This way, existing "qcom,dsi-select-
sec-clocks" property can be used as clock source for the
secondary display when DPU0 is faulty/disabled.

Change-Id: I3ced083b6700921549e7036722a73d7a4f1f41e5
Signed-off-by: Ayushi Makhija <quic_amakhija@quicinc.com>
Signed-off-by: lnxdisplay <lnxdisplay@localhost>