ARM: dts: qcom: Add PMIC PON log device for tuna
Add a PMIC PON log parser device which reads the log stored in
PMK8550 SDAM5 and SDAM6.

while at it add a dependency between the gh-watchdog and
pmic-pon-log devices for tuna boards.  This ensures that
during system boot-up, the driver responsible for the
gh-watchdog device initializes before the one
responsible for the pmic-pon-log device.

Change-Id: I5e2c13807570e5a2e1740e9160cce626fa9004c6
Signed-off-by: Kavya Nunna <quic_knunna@quicinc.com>