ARM: dts: msm: Enable tri-led node for Ravelin
Add compatible string to led DT node for Ravelin. This
was removed from the bulk DT porting for Ravelin on qcom-6.6
device-tree branch.

Also, the pwm driver probe fails when the "qcom,num-lpg-channels"
value is higher than the number of configured channels in the
dtsi for the pwm node. This is because the driver uses the
lpg_sdam_base for every channel during probe and fails if not
configured correctly in the dtsi. So, change the num-lpg-channels
to 3 since only 3 channels are configured and used.

Change-Id: I9dbe8a2bf97316fad69ff6a5f9db42a19b839901
Signed-off-by: Shilpa Suresh <quic_c_sbsure@quicinc.com>