ARM: dts: msm: Adding all the relevant usb changes in kera
1. The high speed phy is required for USB to support HS
usecases. Add eusb node on tuna which includes the
necessary resources for the eusb phy to work.

2. Adds interconnects on KERA USB.

3. QMP phy is used for SS/SSP usb usecases as well as DP use cases
in a target. This change adds the basic resources required along
with the init sequence for functionality.

4. Adding the necessary audio node providing the necessary
resources of the qmi audio to get probed.

5. The GSI event buffers are required for the various GSI related
usecases which are excercised from the dwc3 glue driver.
Add the number of event buffers along with the register offsets
defined.

6. Adding a 4-byte register entry for tcsr_dyn-en-dis to enable/disable USB
dynamically from dwc3-msm-core.

Change-Id: Ia45a09d8d2a54e29ecd0811e97d71c6dd8eaecee
Signed-off-by: Uttkarsh Aggarwal <quic_uaggarwa@quicinc.com>