ARM: dts: msm: Add wcd_usbss reference to USB on tuna
The USB D+/D- signal lines are first routed through the WCD939x USB
subsystem before connecting to the USB controller on MTP and QRD platform
for tuna. Add a phandle to the former to the USB device node. This will
allow the USB driver to control the D+/D- switches when USB 2.0
functionality is needed.

Change-Id: Ie55271923cc893f72602ecbe156617d651bef66e
Signed-off-by: Uttkarsh Aggarwal <quic_uaggarwa@quicinc.com>