ARM: dts: msm: Add High Speed USB support on Kera
Currently, the node for eusb2_phy0 is defined but not used by controller
which makes the associated resources to be consumed. Therefore if phy
probes doesn't happen the controller goes into core soft reset failure.

Fix this by utilizing the node in the controller. This will call the
phy's probe and hence clocks & regulators will be initialized.

Change-Id: Ica46af01154d5e583e011e4f2d84a80fb0589ac8
Signed-off-by: Udipto Goswami <quic_ugoswami@quicinc.com>