ARM: dts: msm: Add wcd_usbss reference to USB on Kera
The USB D+/D- signal lines are first routed through the WCD939x USB
subsystem before connecting to the USB controller on MTP and QRD
platform for kera. Add a phandle to the former to the USB device
node. This will allow the USB driver to control the D+/D- switches
when functionality is needed.

Change-Id: Ie6e76885785cc57974d52df91297a98f300cf666
Signed-off-by: Udipto Goswami <quic_ugoswami@quicinc.com>