ARM: dts: msm: Update regulator support for tuna
Add RET mode support for L2G/L3G for tuna platforms as per
the sleep setting recommendation.

While at it set init mode as LPM for L3G and L6K
regulators. As clients always vote for 0 load, the
regulator framework will not apply it and the HPM
init-mode will not change, leading to higher power
consumption. So update the LPM for L3G and L6K regulators.

Change-Id: I5b210ac3e9ffee94889c2390becfaa5eb6c235ab
Signed-off-by: Kavya Nunna <quic_knunna@quicinc.com>