ARM: dts: msm: Remove some ocp-notifier supplies from sun HDK platforms
Sun HDK platform is expected to work without the GNSS board (which also
holds PMR735D) being attached.  If the board is not attached,
regulator-ocp-notifier probe may fail as it references some regulators
on PMR735D which do not get probed.

Remove those regulator references from regulator-ocp-notifier as they
are not strictly necessary under it.

Change-Id: Ie6f35a9c6142fe54d7191f9a392074f2e39b2bee
Signed-off-by: Fenglin Wu <quic_fenglinw@quicinc.com>